Move url encoded to separate module
